+++ gcpy/benchmark/modules/benchmark_funcs.py
@@ -41,7 +41,7 @@ def create_total_emissions_table(
         template="Emis{}_",
         refmetdata=None,
         devmetdata=None,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates a table of emissions totals (by sector and by inventory)
@@ -105,7 +105,7 @@ def create_total_emissions_table(
             Default value: None
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
 
     Remarks:
         This method is mainly intended for model benchmarking purposes,
@@ -145,6 +145,8 @@ def create_total_emissions_table(
     # molecular weights), which we will need for unit conversions.
     # This is located in the "data" subfolder of this folder where
     # this benchmark.py file is found.
+    if spcdb_dir is None:
+        raise ValueError("The 'spcdb_dir' argument has not been specified!")
     properties = util.read_config_file(
         os.path.join(
             spcdb_dir,
@@ -367,7 +369,7 @@ def create_global_mass_table(
         trop_only=False,
         outfilename="GlobalMass_TropStrat.txt",
         verbose=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates a table of global masses for a list of species in contained in
@@ -410,7 +412,7 @@ def create_global_mass_table(
             Default value: False
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
 
     Remarks:
         This method is mainly intended for model benchmarking purposes,
@@ -435,6 +437,8 @@ def create_global_mass_table(
     # Load a YAML file containing species properties (such as
     # molecular weights), which we will need for unit conversions.
     # This is located in the "data" subfolder of this current directory.2
+    if spcdb_dir is None:
+        raise ValueError("The 'spcdb_dir' argument has not been specified!")
     properties = util.read_config_file(
         os.path.join(
             spcdb_dir,
@@ -599,7 +603,7 @@ def create_mass_accumulation_table(
         trop_only=False,
         outfilename="GlobalMassAccum_TropStrat.txt",
         verbose=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates a table of global mass accumulation for a list of species in
@@ -650,7 +654,7 @@ def create_mass_accumulation_table(
             Default value: False
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
 
     Remarks:
         This method is mainly intended for model benchmarking purposes,
@@ -677,6 +681,9 @@ def create_mass_accumulation_table(
     # Load a YAML file containing species properties (such as
     # molecular weights), which we will need for unit conversions.
     # This is located in the "data" subfolder of this current directory.2
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        raise ValueError("The 'spcdb_dir' argument has not been specified!")
     properties = util.read_config_file(
         os.path.join(
             spcdb_dir,
@@ -894,7 +901,7 @@ def make_benchmark_conc_plots(
         second_ref=None,
         second_dev=None,
         time_mean=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates PDF files containing plots of species concentration
@@ -998,7 +1005,7 @@ def make_benchmark_conc_plots(
             Default value: None
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
         time_mean : bool
             Determines if we should average the datasets over time
             Default value: False
@@ -1010,6 +1017,10 @@ def make_benchmark_conc_plots(
     # ==================================================================
     # Initialization and data read
     # ==================================================================
+
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        raise ValueError("The 'spcdb_dir' argument has not been specified!")
 
     # Create the destination folder
     util.make_directory(dst, overwrite)
@@ -1500,7 +1511,7 @@ def make_benchmark_conc_plots(
         for _, filecat in enumerate(catdict):
             results.append(createplots(filecat))
     # --------------------------------------------
-    
+
     dict_sfc = {list(result.keys())[0]: result[list(
         result.keys())[0]]['sfc'] for result in results}
     dict_500 = {list(result.keys())[0]: result[list(
@@ -1576,7 +1587,7 @@ def make_benchmark_emis_plots(
         weightsdir='.',
         n_job=-1,
         time_mean=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates PDF files containing plots of emissions for model
@@ -1662,7 +1673,7 @@ def make_benchmark_emis_plots(
             Default value: -1
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
         time_mean : bool
             Determines if we should average the datasets over time
             Default value: False
@@ -1678,6 +1689,10 @@ def make_benchmark_emis_plots(
     # =================================================================
     # Initialization and data read
     # =================================================================
+
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        raise ValueError("The 'spcdb_dir' argument has not been specified!")
 
     # Create the destination folder
     util.make_directory(dst, overwrite)
@@ -2011,7 +2026,7 @@ def make_benchmark_emis_tables(
         overwrite=False,
         ref_interval=[2678400.0],
         dev_interval=[2678400.0],
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates a text file containing emission totals by species and
@@ -2069,6 +2084,11 @@ def make_benchmark_emis_tables(
     # ==================================================================
     # Initialization
     # ==================================================================
+
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        msg = "The 'spcdb_dir' argument has not been specified!"
+        raise ValueError(msg)
 
     # Create the destination folder
     util.make_directory(dst, overwrite)
@@ -2204,7 +2224,7 @@ def make_benchmark_jvalue_plots(
         weightsdir='.',
         n_job=-1,
         time_mean=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates PDF files containing plots of J-values for model
@@ -2290,7 +2310,7 @@ def make_benchmark_jvalue_plots(
             Default value: -1
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
         time_mean : bool
             Determines if we should average the datasets over time
             Default value: False
@@ -2314,6 +2334,11 @@ def make_benchmark_jvalue_plots(
     # ==================================================================
     # Initialization
     # ==================================================================
+
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        msg = "The 'spcdb_dir' argument has not been specified!"
+        raise ValueError(msg)
 
     # Create the directory for output
     util.make_directory(dst, overwrite)
@@ -2605,7 +2630,7 @@ def make_benchmark_aod_plots(
         weightsdir='.',
         n_job=-1,
         time_mean=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates PDF files containing plots of column aerosol optical
@@ -2671,7 +2696,7 @@ def make_benchmark_aod_plots(
             Default value: -1
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
         time_mean : bool
             Determines if we should average the datasets over time
             Default value: False
@@ -2679,6 +2704,11 @@ def make_benchmark_aod_plots(
     # ==================================================================
     # Initialization and also read data
     # ==================================================================
+
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        msg = "The 'spcdb_dir' argument has not been specified!"
+        raise ValueError(msg)
 
     # Create destination plots directory
     util.make_directory(dst, overwrite)
@@ -2920,7 +2950,7 @@ def make_benchmark_mass_tables(
         overwrite=False,
         verbose=False,
         label="at end of simulation",
-        spcdb_dir=os.path.dirname(__file__),
+        spcdb_dir=None,
         ref_met_extra=None,
         dev_met_extra=None
 ):
@@ -2969,7 +2999,7 @@ def make_benchmark_mass_tables(
             Default value: False.
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
         ref_met_extra: str
             Path to ref Met file containing area data for use with restart files
             which do not contain the Area variable.
@@ -2979,6 +3009,10 @@ def make_benchmark_mass_tables(
             which do not contain the Area variable.
             Default value: ''
     """
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        msg = "The 'spcdb_dir' argument has not been specified!"
+        raise ValueError(msg)
 
     # ==================================================================
     # Define destination directory
@@ -3181,7 +3215,7 @@ def make_benchmark_mass_accumulation_tables(
         overwrite=False,
         verbose=False,
         label="at end of simulation",
-        spcdb_dir=os.path.dirname(__file__),
+        spcdb_dir=None,
 ):
     """
     Creates a text file containing global mass totals by species and
@@ -3239,8 +3273,12 @@ def make_benchmark_mass_accumulation_tables(
             Default value: False.
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
     """
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        msg = "The 'spcdb_dir' argument has not been specified!"
+        raise ValueError(msg)
 
     # ==================================================================
     # Define destination directory
@@ -3707,7 +3745,7 @@ def make_benchmark_wetdep_plots(
         weightsdir='.',
         n_job=-1,
         time_mean=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Creates PDF files containing plots of species concentration
@@ -3771,11 +3809,15 @@ def make_benchmark_wetdep_plots(
             Default value: -1
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
         time_mean : bool
             Determines if we should average the datasets over time
             Default value: False
     """
+    # Make sure the species database folder is passed
+    if spcdb_dir is None:
+        msg = "The 'spcdb_dir' argument has not been specified!"
+        raise ValueError(msg)
 
     # Create destination plot directory
     util.make_directory(dst, overwrite)
@@ -3971,7 +4013,7 @@ def make_benchmark_aerosol_tables(
         dst='./benchmark',
         overwrite=False,
         is_gchp=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Compute FullChemBenchmark aerosol budgets & burdens
@@ -4004,10 +4046,9 @@ def make_benchmark_aerosol_tables(
             Default value: False
         spcdb_dir: str
             Directory of species_datbase.yml file
-            Default value: Directory of GCPy code repository
+            Default value: None
 
     """
-
     # Create destination directory
     util.make_directory(dst, overwrite)
 
@@ -4015,6 +4056,8 @@ def make_benchmark_aerosol_tables(
     species_list = ["BCPI", "OCPI", "SO4", "DST1", "SALA", "SALC"]
 
     # Read the species database
+    if spcdb_dir is None:
+        raise ValueError("The 'spcdb_dir' argument has not been specified!")
     spcdb = util.read_config_file(
         os.path.join(
             spcdb_dir,
@@ -4270,7 +4313,7 @@ def make_benchmark_operations_budget(
         species=None,
         overwrite=True,
         verbose=False,
-        spcdb_dir=os.path.dirname(__file__)
+        spcdb_dir=None,
 ):
     """
     Prints the "operations budget" (i.e. change in mass after
@@ -4341,6 +4384,9 @@ def make_benchmark_operations_budget(
             Set this switch to True if you wish to print out extra
             informational messages.
             Default value: False
+        spcdb_dir : str
+            Directory containing the species_database.yml file.
+            Default value: None
     """
 
     # ------------------------------------------
@@ -4746,6 +4792,8 @@ def make_benchmark_operations_budget(
 
         # Load a YAML file containing species properties (such as
         # molecular weights), which we will need for unit conversions.
+        if spcdb_dir is None:
+            raise ValueError("The 'spcdb_dir' argument has not been specified!")
         properties = util.read_config_file(
             os.path.join(
                 spcdb_dir,